Address key errors and move modals
